Don't set value to NaN
When clearing the field, we'd end up parsing the empty string as an Int, which
results in `NaN`.

This has the side effect of clearing the field when a non-numeric value has been
entered. Previously the field value would be replaced with `NaN` which makes
sense to pretty much no one.